separating logics into hooks
