+++ components/SendXStream.js
@@ -1,31 +1,14 @@
 import dayjs from "dayjs";
 import React, { useEffect, useState } from "react";
-import { ethers } from "ethers";
-import originAbi from "../data/originAbi.json";
-import destinationAbi from "../data/destinationAbi.json";
 import DatePicker from "./DatePicker";
 import DropSelect from "./DropSelect";
 import StreamInfo from "./StreamInfo";
-import { parseEther } from "ethers/lib/utils.js";
 import { useAccount } from "wagmi";
-import { Framework } from "@superfluid-finance/sdk-core";
-import { toast } from "react-toastify";
-import TestTokenAbi from "../data/TestTokenAbi.json";
 import { getNetwork, fetchBalance } from "@wagmi/core";
 import { bridgeDataConfig } from "@/data/config";
-
 import { ConnextIcon, GoerliIcon, PolygonIcon } from "./icons";
-
 import FlowRateModal from "./FLowrateModal";
-
-const options = [
-  { name: "Wade Cooper" },
-  { name: "Arlene Mccoy" },
-  { name: "Devon Webb" },
-  { name: "Tom Cook" },
-  { name: "Tanya Fox" },
-  { name: "Hellen Schmidt" },
-];
+import useXStream from "@/hooks/xStream/useXStream";
 
 const coins = [
   // { id: "0x07865c6E87B9F70255377e024ace6630C1Eaa37F", name: 'USDC' },
@@ -68,128 +51,8 @@ const SendXStream = () => {
 
   const { chain } = getNetwork();
 
-  const sendStreamSameChain = async () => {
-    const senderAddress = address;
-    const receiverAddress = receipient;
-    const flowRate = amount;
-
-    try {
-      const { ethereum } = window;
-      if (ethereum) {
-        const provider = new ethers.providers.Web3Provider(ethereum);
-        const signer = provider.getSigner();
-        const account = await signer.getAddress();
-
-        const sf = await Framework.create({
-          chainId: 5,
-          provider: provider,
-        });
-
-        const TOKENxContract = await sf.loadSuperToken(
-          "0x3427910EBBdABAD8e02823DFe05D34a65564b1a0"
-        );
-        const TOKENx = TOKENxContract.address;
-
-        try {
-          const createFlowOperation = sf.cfaV1.createFlowByOperator({
-            sender: senderAddress,
-            receiver: receiverAddress,
-            flowRate: flowRate,
-            superToken: TOKENx,
-          });
-
-          console.log("Creating your stream...");
-          toast.info("Creating your stream...");
-
-          const result = await createFlowOperation.exec(signer);
-          console.log(result);
-
-          console.log(`Congrats - you've just created a money stream!`);
-        } catch (error) {
-          console.log(
-            "Hmmm, your transaction threw an error. Make sure that this stream does not already exist, and that you've entered a valid Ethereum address!"
-          );
-          toast.error(
-            "Hmmm, your transaction threw an error. Make sure that this stream does not already exist, and that you've entered a valid Ethereum address!"
-          );
-          console.error(error);
-        }
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  const calculateFlowRate = (amountInEther) => {
-    const now = new Date();
-
-    const endD = new Date(endDate);
-    const timeDiff = Math.abs(endD.getTime() - now.getTime());
-    console.log(timeDiff);
-
-    const amount = ethers.utils.parseEther(amountInEther.toString());
-    const calculatedFlowRate = Math.floor(amount / timeDiff);
-
-    // alert(calculatedFlowRate);
-
-    return calculatedFlowRate;
-  };
-
-  const sendStreamDifferentChain = async () => {
-    try {
-      if (typeof window.ethereum !== "undefined") {
-        const provider = new ethers.providers.Web3Provider(window.ethereum);
-        const signer = provider.getSigner();
-
-        let tokenContract;
-
-        // currently on Goerli Testnet
-        tokenContract = new ethers.Contract(
-          token?.address,
-          TestTokenAbi,
-          signer
-        );
-
-        try {
-          let txn = await tokenContract.approve(
-            bridgeDataConfig[chain.id].xstreamContractAddress,
-            parseEther(amount)
-          );
-          await txn.wait();
-        } catch (error) {
-          console.log("Error in approving tokens ", error);
-          return;
-        }
-
-        const flowRate = calculateFlowRate(amount);
-        const contractAbi = chain?.id == 5 ? originAbi : destinationAbi;
-        const originContract = new ethers.Contract(
-          bridgeDataConfig[chain.id].xstreamContractAddress,
-          contractAbi,
-          signer
-        );
-
-        toast.info("Creating your XStream...");
-        const transaction = await originContract._sendFlowMessage(                //_sendFlowMessage
-            "1",                    //streamActionType
-            receipient,             //receiver
-            flowRate,              //flowRate
-            "80000000000000000",    //relayer fees
-            "300",                  //slippage
-            parseEther(amount),     //amount of tokens to send
-            token?.address,
-            bridgeDataConfig[toChain?.id].xstreamContractAddress,
-            bridgeDataConfig[toChain?.id].connextDomainId,
-            { value: parseEther("0.08") }
-        )
-        await transaction.wait();
-        toast.info("Transaction Submitted...");
-      }
-    } catch (error) {
-      console.error(error);
-      toast.error(error.message);
-    }
-  };
+  // custom hooks
+  const hookXStream = useXStream();
 
   const handleSubmit = async (e) => {
     e.preventDefault();
@@ -212,34 +75,18 @@ const SendXStream = () => {
       // alert(toChain.name, fromChain.name);
 
       if (toChain.name === fromChain.name) {
-        await sendStreamSameChain();
+        await hookXStream.sendStreamSameChain();
       } else if (toChain.name !== fromChain.name) {
-        await sendStreamDifferentChain();
+        await hookXStream.sendStreamDifferentChain();
       }
     } catch (error) {
       console.error(error);
     }
   };
 
-  const getBalance = async () => {
-    try {
-      const { ethereum } = window;
-      if (ethereum) {
-        const tokenBalance = await fetchBalance({
-            address: address,
-            token: token?.address
-        });
-        console.log("the current selected tokenbalance is ", tokenBalance);
-        setBalance(tokenBalance.formatted);
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
   useEffect(() => {
     if (token?.address) {
-        getBalance();
+        hookXStream.getBalance();
     }
   }, [token?.address, chain?.id]);
